# openapi.yaml
openapi: 3.0.3
info:
  title: GUARDIAN API
  description: |
    GUARDIAN Pharmaceutical Compliance Analysis API
    
    A comprehensive REST API for pharmaceutical protocol compliance analysis against European Pharmacopoeia standards.
    The system provides document processing, vector similarity search, LLM-powered compliance analysis, and professional report generation.
    
    ## Features
    
    - **Protocol Analysis**: AI-powered compliance analysis against European Pharmacopoeia standards
    - **Document Management**: Upload, process, and manage reference documents
    - **Vector Search**: Similarity search across pharmacopoeia sections
    - **Report Generation**: Professional PDF and HTML reports with visualizations
    - **Batch Processing**: Handle multiple protocols and generate bulk reports
    
    ## Authentication
    
    Currently, the API operates without authentication in development mode. 
    Production deployments should implement appropriate authentication mechanisms.
    
    ## Rate Limiting
    
    API endpoints are rate-limited to ensure system stability:
    - Standard endpoints: 100 requests/minute
    - Analysis endpoints: 10 requests/minute
    - Upload endpoints: 20 requests/minute
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns detailed error information in JSON format.
    
  version: 1.0.0
  contact:
    name: GUARDIAN Support
    email: guardian-support@qmul.ac.uk
...
  license:
    name: MIT
...
servers:
...
paths:
...
  /api/search:
    post:
      tags: [Search]
      summary: Search pharmacopoeia sections
      description: Performs vector similarity search across indexed pharmacopoeia sections
      operationId: searchSections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /api/search/suggestions:
    get:
      tags: [Search]
      summary: Get search suggestions
      description: Provides search query suggestions based on indexed content
      operationId: getSearchSuggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        '200':
          description: Search suggestions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchSuggestionsResponse'
  /api/search/multi-index:
    post:
      tags: [Search]
      summary: Multi-index search
      description: Searches across multiple vector indices simultaneously
      operationId: multiIndexSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiIndexSearchRequest'
      responses:
        '200':
          description: Multi-index search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiIndexSearchResponse'
...
components:
  schemas:
...
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Search query text
          minLength: 2
          maxLength: 1000
        top_k:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Number of results to return
        similarity_threshold:
          type: number
          minimum: 0.0
          maximum: 1.0
          default: 0.5
        filters:
          type: object
          description: Search filters
          properties:
            document_type:
              type: string
            chapter:
              type: string
    SearchResponse:
      type: object
      required:
        - results
        - query
        - total_results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        query:
          type: string
        total_results:
          type: integer
        search_time:
          type: number
          description: Search time in seconds
    SearchResult:
      type: object
      required:
        - id
        - similarity_score
        - content
      properties:
        id:
          type: string
        similarity_score:
          type: number
        content:
          type: string
        metadata:
          type: object
        document_id:
          type: string
    SearchSuggestionsResponse:
      type: object
      required:
        - suggestions
      properties:
        suggestions:
          type: array
          items:
            type: string
        query:
          type: string
    MultiIndexSearchRequest:
      type: object
      required:
        - query
        - indices
      properties:
        query:
          type: string
        indices:
          type: array
          items:
            type: string
        top_k:
          type: integer
          default: 10
    MultiIndexSearchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/SearchResult'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - ApiKeyAuth: []
  - BearerAuth: []